# gen/openapiv2/loom/v1/loom.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "loom/v1/loom.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/mcp/servers": {
      "get": {
        "summary": "ListMCPServers lists all configured MCP servers.",
        "operationId": "LoomService_ListMCPServers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListMCPServersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "LoomService"
        ]
      },
      "post": {
        "summary": "AddMCPServer adds a new MCP server configuration.",
        "operationId": "LoomService_AddMCPServer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AddMCPServerResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "AddMCPServerRequest adds a new MCP server.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AddMCPServerRequest"
            }
          }
        ],
        "tags": [
          "LoomService"
        ]
      }
    },
    "/v1/mcp/servers/{serverName}": {
      "get": {
        "summary": "GetMCPServer retrieves a specific MCP server configuration.",
        "operationId": "LoomService_GetMCPServer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1MCPServerInfo"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "serverName",
            "description": "Server name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "LoomService"
        ]
      },
      "delete": {
        "summary": "DeleteMCPServer removes an MCP server configuration.",
        "operationId": "LoomService_DeleteMCPServer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteMCPServerResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "serverName",
            "description": "Server name to delete",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "force",
            "description": "Force delete even if server is running",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "tags": [
          "LoomService"
        ]
      },
      "put": {
        "summary": "UpdateMCPServer updates an existing MCP server configuration.",
        "operationId": "LoomService_UpdateMCPServer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1MCPServerInfo"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "serverName",
            "description": "Server name to update",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoomServiceUpdateMCPServerBody"
            }
          }
        ],
        "tags": [
          "LoomService"
        ]
      }
    },
...
    "/v1/mcp/servers:test": {
      "post": {
        "summary": "TestMCPServerConnection tests an MCP server configuration without persisting it.\nThis allows users to validate their configuration (command, args, env vars) before saving.",
        "operationId": "LoomService_TestMCPServerConnection",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1TestMCPServerConnectionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "TestMCPServerConnectionRequest tests an MCP server configuration.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1TestMCPServerConnectionRequest"
            }
          }
        ],
        "tags": [
          "LoomService"
        ]
      }
    },
...
    "/v1/weave:stream": {
      "post": {
        "summary": "StreamWeave executes a query and streams incremental progress updates.",
        "operationId": "LoomService_StreamWeave",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1WeaveProgress"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1WeaveProgress"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "WeaveRequest initiates a query execution.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1WeaveRequest"
            }
          }
        ],
        "tags": [
          "LoomService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "LoomServiceUpdateMCPServerBody": {
      "type": "object",
      "properties": {
        "enabled": {
          "type": "boolean",
          "title": "Whether to enable the server"
        },
        "transport": {
          "type": "string",
          "title": "Transport type (stdio, http, sse)"
        },
        "command": {
          "type": "string",
          "title": "Command to execute"
        },
        "args": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Command arguments"
        },
        "env": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Environment variables"
        },
        "workingDir": {
          "type": "string",
          "title": "Working directory"
        },
        "restartIfRunning": {
          "type": "boolean",
          "title": "Whether to restart the server if it's currently running"
        },
        "toolFilter": {
          "$ref": "#/definitions/v1ToolFilterConfig",
          "title": "Tool filter configuration"
        },
        "timeoutSeconds": {
          "type": "integer",
          "format": "int32",
          "title": "Timeout in seconds for restart operation (default: 1 second)"
        }
      },
      "description": "UpdateMCPServerRequest updates an MCP server."
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AddMCPServerRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Server name (unique identifier)"
        },
        "enabled": {
          "type": "boolean",
          "title": "Whether to enable the server"
        },
        "transport": {
          "type": "string",
          "title": "Transport type (stdio, http, sse)"
        },
        "command": {
          "type": "string",
          "title": "Command to execute"
        },
        "args": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Command arguments"
        },
        "env": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Environment variables"
        },
        "workingDir": {
          "type": "string",
          "title": "Working directory"
        },
        "autoStart": {
          "type": "boolean",
          "title": "Whether to start the server immediately after adding"
        },
        "toolFilter": {
          "$ref": "#/definitions/v1ToolFilterConfig",
          "title": "Tool filter configuration"
        }
      },
      "description": "AddMCPServerRequest adds a new MCP server."
    },
    "v1AddMCPServerResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "title": "Whether addition was successful"
        },
        "message": {
          "type": "string",
          "title": "Status message"
        },
        "server": {
          "$ref": "#/definitions/v1MCPServerInfo",
          "title": "Server information"
        }
      },
      "description": "AddMCPServerResponse confirms MCP server addition."
    },
...
    "v1CostInfo": {
      "type": "object",
      "properties": {
        "totalCostUsd": {
          "type": "number",
          "format": "double",
          "title": "Total cost (USD)"
        },
        "llmCost": {
          "$ref": "#/definitions/v1LLMCost",
          "title": "LLM cost breakdown"
        },
        "backendCostUsd": {
          "type": "number",
          "format": "double",
          "title": "Backend execution cost (if applicable)"
        }
      },
      "description": "CostInfo tracks LLM and execution costs."
    },
...
    "v1DataReference": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "Unique identifier for this data chunk"
        },
        "sizeBytes": {
          "type": "string",
          "format": "int64",
          "title": "Size in bytes"
        },
        "location": {
          "$ref": "#/definitions/v1StorageLocation",
          "title": "Where the data is stored"
        },
        "checksum": {
          "type": "string",
          "title": "Checksum for validation (SHA-256)"
        },
        "compressed": {
          "type": "boolean",
          "title": "Whether data is compressed"
        },
        "contentType": {
          "type": "string",
          "title": "MIME type or data format hint"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Metadata about the data"
        },
        "storedAt": {
          "type": "string",
          "format": "int64",
          "title": "Timestamp when data was stored (Unix milliseconds)"
        }
      },
      "description": "DataReference points to data in shared memory."
    },
...
    "v1DeleteMCPServerResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "title": "Whether deletion was successful"
        },
        "message": {
          "type": "string",
          "title": "Status message"
        }
      },
      "description": "DeleteMCPServerResponse confirms deletion."
    },
...
    "v1ExecutionResult": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "title": "Result type (query, operation, document, etc.)"
        },
        "dataJson": {
          "type": "string",
          "title": "Structured data (JSON) - for small results only"
        },
        "rowCount": {
          "type": "integer",
          "format": "int32",
          "title": "Row count (for query results)"
        },
        "columns": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Columns (for tabular results)"
        },
        "bytesScanned": {
          "type": "string",
          "format": "int64",
          "title": "Data scanned (bytes)"
        },
        "durationMs": {
          "type": "string",
          "format": "int64",
          "title": "Execution duration (milliseconds)"
        },
        "backendMetadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Backend-specific metadata"
        },
        "dataReference": {
          "$ref": "#/definitions/v1DataReference",
          "title": "Data reference for large results (stored in shared memory)"
        }
      },
      "description": "ExecutionResult contains domain-specific execution results."
    },
    "v1ExecutionStage": {
      "type": "string",
      "enum": [
        "EXECUTION_STAGE_UNSPECIFIED",
        "EXECUTION_STAGE_PATTERN_SELECTION",
        "EXECUTION_STAGE_SCHEMA_DISCOVERY",
        "EXECUTION_STAGE_LLM_GENERATION",
        "EXECUTION_STAGE_TOOL_EXECUTION",
        "EXECUTION_STAGE_HUMAN_IN_THE_LOOP",
        "EXECUTION_STAGE_GUARDRAIL_CHECK",
        "EXECUTION_STAGE_SELF_CORRECTION",
        "EXECUTION_STAGE_COMPLETED",
        "EXECUTION_STAGE_FAILED"
      ],
      "default": "EXECUTION_STAGE_UNSPECIFIED",
      "description": "ExecutionStage enum for tracking progress.\n\n - EXECUTION_STAGE_HUMAN_IN_THE_LOOP: Waiting for human response"
    },
...
    "v1HITLRequestInfo": {
      "type": "object",
      "properties": {
        "requestId": {
          "type": "string",
          "title": "Request ID (generated by contact_human tool or clarification system)"
        },
        "question": {
          "type": "string",
          "title": "Question being asked to the human"
        },
        "requestType": {
          "type": "string",
          "title": "Request type (approval, decision, input, review, clarification)"
        },
        "priority": {
          "type": "string",
          "title": "Priority (low, normal, high, critical)"
        },
        "timeoutSeconds": {
          "type": "integer",
          "format": "int32",
          "title": "Timeout in seconds"
        },
        "contextJson": {
          "type": "string",
          "title": "Additional context as JSON"
        },
        "questionId": {
          "type": "string",
          "title": "Question ID for clarification questions (used with AnswerClarificationQuestion RPC)"
        },
        "options": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Suggested options for the user (if applicable)"
        }
      },
      "description": "HITLRequestInfo carries information about a human-in-the-loop request."
    },
...
    "v1LLMCost": {
      "type": "object",
      "properties": {
        "totalTokens": {
          "type": "integer",
          "format": "int32",
          "title": "Total tokens used"
        },
        "inputTokens": {
          "type": "integer",
          "format": "int32",
          "title": "Input tokens"
        },
        "outputTokens": {
          "type": "integer",
          "format": "int32",
          "title": "Output tokens"
        },
        "costUsd": {
          "type": "number",
          "format": "double",
          "title": "Cost for this LLM usage (USD)"
        },
        "model": {
          "type": "string",
          "title": "Model used"
        },
        "provider": {
          "type": "string",
          "title": "Provider"
        }
      },
      "description": "LLMCost tracks token usage and costs."
    },
...
    "v1ListMCPServersResponse": {
      "type": "object",
      "properties": {
        "servers": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1MCPServerInfo"
          },
          "title": "List of MCP servers"
        },
        "totalCount": {
          "type": "integer",
          "format": "int32",
          "title": "Total count"
        }
      },
      "description": "ListMCPServersResponse returns MCP servers."
    },
...
    "v1MCPServerInfo": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Server name"
        },
        "enabled": {
          "type": "boolean",
          "title": "Whether server is enabled"
        },
        "connected": {
          "type": "boolean",
          "title": "Whether server is currently connected"
        },
        "transport": {
          "type": "string",
          "title": "Transport type (stdio, http, sse)"
        },
        "command": {
          "type": "string",
          "title": "Command to execute"
        },
        "args": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Command arguments"
        },
        "env": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Environment variables"
        },
        "status": {
          "type": "string",
          "title": "Server status (running, stopped, error, disabled)"
        },
        "error": {
          "type": "string",
          "title": "Error message if status is error"
        },
        "uptimeSeconds": {
          "type": "string",
          "format": "int64",
          "title": "Uptime in seconds"
        },
        "toolCount": {
          "type": "integer",
          "format": "int32",
          "title": "Number of tools exposed by this server"
        }
      },
      "description": "MCPServerInfo provides information about an MCP server."
    },
...
    "v1StorageLocation": {
      "type": "string",
      "enum": [
        "STORAGE_LOCATION_UNSPECIFIED",
        "STORAGE_LOCATION_MEMORY",
        "STORAGE_LOCATION_DISK",
        "STORAGE_LOCATION_DATABASE"
      ],
      "default": "STORAGE_LOCATION_UNSPECIFIED",
      "description": "StorageLocation indicates where data is stored.\n\n - STORAGE_LOCATION_MEMORY: In-memory shared storage\n - STORAGE_LOCATION_DISK: Disk overflow cache\n - STORAGE_LOCATION_DATABASE: Queryable SQL storage"
    },
...
    "v1TestMCPServerConnectionRequest": {
      "type": "object",
      "properties": {
        "transport": {
          "type": "string",
          "title": "Transport type (stdio, http, sse)"
        },
        "command": {
          "type": "string",
          "title": "Command to execute"
        },
        "args": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Command arguments"
        },
        "env": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Environment variables"
        },
        "workingDir": {
          "type": "string",
          "title": "Working directory"
        },
        "toolFilter": {
          "$ref": "#/definitions/v1ToolFilterConfig",
          "title": "Tool filter configuration"
        },
        "timeoutSeconds": {
          "type": "integer",
          "format": "int32",
          "title": "Timeout for connection test in seconds (default: 10)"
        }
      },
      "description": "TestMCPServerConnectionRequest tests an MCP server configuration."
    },
    "v1TestMCPServerConnectionResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "title": "Whether the connection test succeeded"
        },
        "message": {
          "type": "string",
          "title": "Status message"
        },
        "error": {
          "type": "string",
          "title": "Error details if failed"
        },
        "toolCount": {
          "type": "integer",
          "format": "int32",
          "title": "Number of tools discovered (if successful)"
        },
        "capabilities": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Server capabilities (if successful)"
        },
        "latencyMs": {
          "type": "string",
          "format": "int64",
          "title": "Connection latency in milliseconds"
        }
      },
      "description": "TestMCPServerConnectionResponse returns test results."
    },
...
    "v1ToolFilterConfig": {
      "type": "object",
      "properties": {
        "all": {
          "type": "boolean",
          "title": "Register all tools"
        },
        "include": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Include only these tools"
        },
        "exclude": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Exclude these tools"
        }
      },
      "description": "ToolFilterConfig configures which tools to register from an MCP server."
    },
...
    "v1WeaveProgress": {
      "type": "object",
      "properties": {
        "stage": {
          "$ref": "#/definitions/v1ExecutionStage",
          "title": "Current stage"
        },
        "progress": {
          "type": "integer",
          "format": "int32",
          "title": "Progress percentage (0-100)"
        },
        "message": {
          "type": "string",
          "title": "Current operation description"
        },
        "toolName": {
          "type": "string",
          "title": "Tool being executed (if applicable)"
        },
        "timestamp": {
          "type": "string",
          "format": "int64",
          "title": "Timestamp"
        },
        "partialResult": {
          "$ref": "#/definitions/v1ExecutionResult",
          "title": "Partial result (available at certain stages)"
        },
        "hitlRequest": {
          "$ref": "#/definitions/v1HITLRequestInfo",
          "title": "HITL request info (only when stage == EXECUTION_STAGE_HUMAN_IN_THE_LOOP)"
        },
        "partialContent": {
          "type": "string",
          "title": "Token streaming fields\nAccumulated content so far (for streaming responses)"
        },
        "isTokenStream": {
          "type": "boolean",
          "title": "True if this is a token streaming update"
        },
        "tokenCount": {
          "type": "integer",
          "format": "int32",
          "title": "Running token count (for streaming responses)"
        },
        "ttftMs": {
          "type": "string",
          "format": "int64",
          "title": "Time to first token in milliseconds (for streaming responses)"
        },
        "cost": {
          "$ref": "#/definitions/v1CostInfo",
          "title": "Cost information (included in completion event)"
        }
      },
      "description": "WeaveProgress streams incremental updates during execution."
    },
    "v1WeaveRequest": {
      "type": "object",
      "properties": {
        "query": {
          "type": "string",
          "title": "User's natural language query"
        },
        "sessionId": {
          "type": "string",
          "title": "Session ID for conversation continuity (optional)"
        },
        "backendConfig": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Backend configuration overrides (optional)"
        },
        "maxRounds": {
          "type": "integer",
          "format": "int32",
          "title": "Maximum execution rounds (default: 3)"
        },
        "timeoutSeconds": {
          "type": "integer",
          "format": "int32",
          "title": "Timeout in seconds (default: 300)"
        },
        "context": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Context variables for prompt interpolation"
        },
        "forcePattern": {
          "type": "string",
          "title": "Force specific pattern (bypass pattern selection)"
        },
        "enableTrace": {
          "type": "boolean",
          "title": "Enable tracing for this request"
        },
        "agentId": {
          "type": "string",
          "title": "Agent ID to route request to (optional, uses default if not specified)"
        }
      },
      "description": "WeaveRequest initiates a query execution."
    },
...
  }
}